livekit-agent-backend/requirements.txt
python-dotenv
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
